[#14491] robot: Add option not to record keyword call.
This change allows us to create keywords that will not be replayed with
`Provides/Requires` keywords.

It is useful when issuing commands that should be executed only once (e.g.
affecting the host system) or when the command should be local to a test
and does not influence the state itself (e.g. setting up logs).