[#12743] VerilatorIntLib: Fix compiler warnings
Including <string> in Socket.h resolves AppleClang error (g++ doesn't
even emit a warning). Including it might have been the intention of the
original author as it contained duplicated line including <string.h>.

On Windows, setsockopt function returns SOCKET_ERROR in case of any
problem, not INVALID_SOCKET.

Created peripheral is also prevented from logging to the terminal (which
is the same as Renode's). This is achieved by passing SettingsFlag without
ENABLE_LOG (which is the default) while creating CTCPClient.