[#12743] VerilatorPlugin: Pause thread handling async communication
ReceiveLoop is a function running inside independent thread.

Normally, pause is not a problem as most methods are called by tlibs.
Here however, asyncEvents-communication thread is running on it's own so
it has to be paused manually as there should be no communication during
emulation pause.

Additionally, now the communication thread starts with the emulation.
Previously it was started right after establishing connection with the
verilated peripheral.