[#18660] conda-package: Prevent searching for 'libdl.so.2' on macOS
Searching for 'libdl.so.2' and creating it's link was primarily intended
for Ubuntu but it doesn't make a problem for any GNU/Linux, as they all
contain that library.

It causes problems on macOS, however, because there is no such library,
so the 'find / ...' command is executed each time conda activates the
environment with the Renode package installed. Therefore, checking
whether the OS is 'Linux' was added as the first test.

The single '=' is intended. Not all shells support '==' in 'test'.