[#18660] conda-package: Whitelist .NET's mscoree.dll
Most of the Renode libraries on Windows need 'mscoree.dll', because it's
the essential .NET library. It is neither packaged with Renode nor any
'requirements:run' package, which triggers multiple warnings during
building.

Whitelisting this library in the 'build:missing_dso_whitelist' simply
informs 'conda-build' that it will be provided from outside Conda.