[#19378] conda-package: Use C compiler from conda package
Tools for building Renode's tlibs are picked according to the properties
set in the Infrastructure submodule's '${OS}-properties.csproj' files.

For Linux and OSX there are full paths so they will be modified by the
patch to use tools conda packages set as the AR and CC variables.

The 'gcc' in 'check_weak_implementations.sh' will be also changed to the
'$CC' as compiler in conda isn't available through the simple 'gcc'.

The patch was created with 'git diff --submodule=diff'.