[#44214] gdb_compare: Fix GDBInstance.task type safety
The expect function from pexpect can return 2 different types depending
on whether async was used or not. It either returns an integer index
(async_ == True) or an awaitable asyncio.coroutine (async_ == False).

To be sure we are allowed to call `await self.task`, it is necessary to
make that distinction and only assign to self.task if async_ == True.

Also added the Awaitable[Any] type hint.