[#44214] gdb_compare: Add detection for dying processes
It seems reasonable to check if a process is alive before messaging it.
Without this change, if a process (Renode, reference, or either GDB)
dies, the script will most likely timeout and it's not immediately clear
what happened. Now, dead processes are detected before executing any
command, and an adequate message is printed, saving some time for the
user.