[#44214] gdb_compare: Add more RE_GDB_ERRORS
These regexes were added to detect passing a completely wrong value as
reference port number (i.e. "abcd", not even a number). Formerly this
would cause an ambiguous timeout.
The scope of these regexes is larger though - they detect general name
and filename errors, as well as attempts to step when the program
(Renode simulation) is not running.